--- src/app/tests/MockAclTestContext.cpp
@@ -1,115 +1,0 @@
-/*
- *    Copyright (c) 2021 Project CHIP Authors
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-#include <app/tests/MockAclTestContext.h>
-
-#include <access/AccessControl.h>
-#include <access/examples/PermissiveAccessControlDelegate.h>
-#include <app/InteractionModelEngine.h>
-#include <lib/support/CodeUtils.h>
-#include <lib/support/ErrorStr.h>
-
-namespace {
-using namespace chip;
-using namespace chip::Access;
-using Entry                     = chip::Access::AccessControl::Entry;
-using EntryIterator             = chip::Access::AccessControl::EntryIterator;
-using Target                    = Entry::Target;
-chip::ClusterId kTestClusterId1 = 6;
-
-class TestDeviceTypeResolver : public AccessControl::DeviceTypeResolver
-{
-public:
-    bool IsDeviceTypeOnEndpoint(DeviceTypeId deviceType, EndpointId endpoint) override { return false; }
-} gDeviceTypeResolver;
-
-AccessControl gTestAccessControl;
-
-class TestAccessControlDelegate : public AccessControl::Delegate
-{
-public:
-    CHIP_ERROR Init() override { return CHIP_NO_ERROR; }
-    void Finish() override {}
-
-    // Capabilities
-    CHIP_ERROR GetMaxEntryCount(size_t & value) const override
-    {
-        value = 0;
-        return CHIP_NO_ERROR;
-    }
-
-    // Actualities
-    CHIP_ERROR GetEntryCount(size_t & value) const override
-    {
-        value = 0;
-        return CHIP_NO_ERROR;
-    }
-
-    // Preparation
-    CHIP_ERROR PrepareEntry(Entry & entry) override { return CHIP_NO_ERROR; }
-
-    // CRUD
-    CHIP_ERROR CreateEntry(size_t * index, const Entry & entry, FabricIndex * fabricIndex) override { return CHIP_NO_ERROR; }
-    CHIP_ERROR ReadEntry(size_t index, Entry & entry, const FabricIndex * fabricIndex) const override { return CHIP_NO_ERROR; }
-    CHIP_ERROR UpdateEntry(size_t index, const Entry & entry, const FabricIndex * fabricIndex) override { return CHIP_NO_ERROR; }
-    CHIP_ERROR DeleteEntry(size_t index, const FabricIndex * fabricIndex) override { return CHIP_NO_ERROR; }
-
-    // Iteration
-    CHIP_ERROR Entries(EntryIterator & iterator, const FabricIndex * fabricIndex) const override { return CHIP_NO_ERROR; }
-
-    // Check
-    CHIP_ERROR Check(const SubjectDescriptor & subjectDescriptor, const RequestPath & requestPath,
-                     Privilege requestPrivilege) override
-    {
-        if (requestPath.cluster == kTestClusterId1)
-        {
-            return CHIP_ERROR_ACCESS_DENIED;
-        }
-        return CHIP_NO_ERROR;
-    }
-};
-
-AccessControl::Delegate * GetTestAccessControlDelegate()
-{
-    static TestAccessControlDelegate accessControlDelegate;
-    return &accessControlDelegate;
-}
-} // namespace
-
-namespace chip {
-namespace Test {
-
-CHIP_ERROR MockAclTestContext::Init()
-{
-    ReturnErrorOnFailure(Super::Init());
-    ReturnErrorOnFailure(chip::app::InteractionModelEngine::GetInstance()->Init(&GetExchangeManager(), &GetFabricTable()));
-
-    Access::SetAccessControl(gTestAccessControl);
-    ReturnErrorOnFailure(Access::GetAccessControl().Init(GetTestAccessControlDelegate(), gDeviceTypeResolver));
-
-    return CHIP_NO_ERROR;
-}
-
-void MockAclTestContext::Shutdown()
-{
-    Access::GetAccessControl().Finish();
-
-    chip::app::InteractionModelEngine::GetInstance()->Shutdown();
-    Super::Shutdown();
-}
-
-} // namespace Test
-} // namespace chip
